Refactor: Replace 64-bit hash usage with 32-bit hash in core components, update `Symbol` constructors for name and hash handling, and adjust related methods and data structures for consistency and efficiency.

+++ src/Lexer/LangSpec.cpp
@@ -115,7 +115,7 @@ void LangSpec::setupCharFlags()
     charFlags_['U'].add(CharFlagsE::Escape);
 }
 
-TokenId LangSpec::keyword(std::string_view name, uint64_t hash) const
+TokenId LangSpec::keyword(std::string_view name, uint32_t hash) const
 {
     const auto ptr = keywordMap_.find(name, hash);
     if (!ptr)
