+++ src/Lexer/LangSpec.h
@@ -44,7 +44,7 @@ public:
     bool isEscape(uint32_t c) const { return c < 256 && charFlags_[c].has(CharFlagsE::Escape); }
     bool isOption(uint32_t c) const { return c < 256 && charFlags_[c].has(CharFlagsE::Option); }
 
-    TokenId keyword(std::string_view name, uint64_t hash) const;
+    TokenId keyword(std::string_view name, uint32_t hash) const;
     TokenId keyword(std::string_view name) const;
 
     static constexpr std::string_view VERIFY_COMMENT_OPTION   = "swc-option";
